fix: Save tracks in undo state instead of DOM element
For some reason, the right-click behavior does not work as intended
after cloning a DOM. To work around this issue, the undo state now
stores the track info instead of the DOM element. Now after undo, the
table is reconstructed and the right-click behavior works after undo.
The downside is that it takes longer to perform undo/redo action, but as
long as the playlist does not contain hundreds of tracks the response
time is acceptable.