feat: Detect no session and send heartbeats to keep session alive
Before, the user would not be properly notified when the session has
timed out due to inactivity. This meant that the user could do a lot of
work that would not be saved. This patch adds a check that if no session
is active then the user is automatically sent to login page. Also, this
patch adds sending of heartbeat in order to keep sessions alive.